fix:remove dup content and reconstructed code
